Create a bigger initial disk on Parallels Desktop 13
That is a workaround for "plain" -> "expanding" disk conversion on Parallels Desktop 13.0.* .
To get it working as before, we need to create a bigger disk image. The minimal size is 1891Mb.
It makes sense only in the creation phase - after that, the disk will be always resized to d.DiskSize